+++ dingtalk/core.py
@@ -6,9 +6,10 @@ import requests
 from hmac import HMAC
 from hashlib import sha256
 from base64 import b64encode
-from .exceptions import DingTlakException
+from .exceptions import DingTalkException
 import time
 import logging
+from urllib.parse import quote
 
 URL = "https://oapi.dingtalk.com"
 
@@ -35,8 +36,8 @@ class Core(object):
         res = requests.get(
             url, {"appkey": self._appkey, "appsecret": self._appsecret}).json()
         # 'errcode': 0, 'access_token': 'f0a2837a3412334589cda9c3ab3a93e6', 'errmsg': 'ok', 'expires_in': 7200
-        if not res:
-            raise Exception()
+        if res['errcode'] != 0:
+            raise DingTalkException(res)
         return res['access_token']
 
     def _get_sso_access_token(self):
@@ -50,9 +51,8 @@ class Core(object):
         res = requests.get(
             url, {'corpid': self._corpid, 'corpsecret': self._appsecret})
         if not res:
-            raise DingTlakException(res)
+            raise DingTalkException(res)
         return res['access_token']
-        
 
     def _get_corp_access_token(self):
         """
@@ -66,26 +66,23 @@ class Core(object):
             "suiteTicket": self._suitticket,
             "auth_corpid": self._corpid,
             "timestamp": timestamp,
-            "signature": self._sign_corp_request(timestamp)
+            "signature": self._sign_corp_request(self._appsecret, timestamp)
         })
         return res
 
-    def _sign_corp_request(self, timestamp):
+    def _sign_corp_request(self, appsecret, timestamp):
         """
         compute signature for third app request.
         """
         signstring = f"{timestamp}\n{self._suitticket}"
-        return b64encode(HMAC(self._appsecret, signstring, sha256).digest())
+        return qoute(b64encode(HMAC(appsecret.encode('utf-8'), signstring, sha256).digest()).decode('utf-8'))
 
     def _post(self, url, data):
-        try:
-            # [FIXME] other two type requests
-            access_token = self._get_enterprise_access_token()
-            res = requests.post(
-                f"{url}?access_token={access_token}", json=data).json()
-            if res['errcode'] != 0:
-                _logger.debug(f"[DingTalk Request]:{data}")
-                raise DingTlakException(**res)
-            return res
-        except Exception as err:
-            raise Exception(err)
+        # [FIXME] other two type requests
+        access_token = self._get_enterprise_access_token()
+        res = requests.post(
+            f"{url}?access_token={access_token}", json=data).json()
+        if res['errcode'] != 0:
+            _logger.debug(f"[DingTalk Request]:{data}")
+            raise DingTalkException(**res)
+        return res