+++ dingtalk/dingtalk.py
@@ -6,9 +6,10 @@ from .core import Core
 from .contact import Department, Role, User
 from .message import Message
 
+
 class DingTalk(object):
 
-    def __init__(self, corpid, appkey, appsecret,agentid,suitticket=None):
+    def __init__(self,  appkey, appsecret, corpid=None, agentid=None, suitticket=None):
         """
         init dingtalk client
 
@@ -22,7 +23,7 @@ class DingTalk(object):
         self._appkey = appkey
         self._appsecret = appsecret
         self._suitticket = suitticket
-        self._agentid  = agentid
+        self._agentid = agentid
 
     core = Core()
     department = Department()