+++ tests/test_core.py
@@ -13,7 +13,7 @@ class TestCore(TestCase):
     def setUpClass(cls):
         cls.appkey = "dingtjjs1pr7nlgmtoxc"
         cls.appsecret = "8a6Ltc8_w-BNpqVOXg3dUH_1PHxxgmWnuf6Gt1ZcQqaMR3fYDDD6rs3Jnmzxr9uy"
-        cls.dingtalk = DingTalk("", cls.appkey, cls.appsecret)  
+        cls.dingtalk = DingTalk(cls.appkey, cls.appsecret)  
 
     def test_get_enterprise_access_token(self):
         res = self.dingtalk.core._get_enterprise_access_token()