+++ changelog.md
@@ -6,3 +6,17 @@
 - Role Management Api
 - User Management Api
 - Message Api
+
+## 0.0.2
+
+- Add App Oauth Api
+
+## 0.0.3
+
+- Fix Exception Error
+- Fix Signature compution.
+- Redefine DingTalk init method.
+
+## 0.0.4
+
+- Fix bugs.